+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed suggestion popup positioning in query editor in some cases
 - Modified `Widget` to utilize deep comparison for page params, ensuring more accurate detection of changes
 - Added `Widget#view.attachTooltip(el, config, data, context)` method
+- Removed `Popup#options` in favor of properties
 
 ## 1.0.0-beta.78 (21-09-2023)